backend/requirement.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
motor==3.7.1
pymongo>=4.9.0,<5
pydantic==2.5.2
python-dotenv==1.0.0
pillow==10.0.0
pytesseract==0.3.10
pdfplumber==0.9.0
pdf2image==1.16.0
python-dateutil==2.8.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
httpx==0.24.1
